moved the loagout functionality to the settings page
