+++ src/Pages/Home.js
@@ -1,21 +1,9 @@
-import { useContext } from "react";
 import styled from "styled-components";
-import { AuthenticationContext } from "../Contexts/AuthenticationContenxt";
 
 function Home() {
-  const { user, dispatch } = useContext(AuthenticationContext);
-
-  const onClickHandler = () => {
-    localStorage.removeItem("user");
-    dispatch({ type: "SIGNOUT" });
-  };
-
   return (
     <HomeContainer>
-      <h1>{`Welcome ${user.email}`}</h1>
-      <StyledButton type="button" onClick={onClickHandler}>
-        Signout
-      </StyledButton>
+      <h1>Home page</h1>
     </HomeContainer>
   );
 }
@@ -29,24 +17,6 @@ const HomeContainer = styled.div`
   align-items: center;
   overflow-y: scroll;
 
-  h1 {
-    margin-bottom: 2rem;
-  }
-`;
-const StyledButton = styled.button`
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  background-color: #6ee7b7;
-  border-radius: 10px;
-  padding: 1rem;
-  border: unset;
-  font-weight: bolder;
-  font-size: 1rem;
-  cursor: pointer;
-  &:hover {
-    background-color: #34d399;
-  }
 `;
 
 export default Home;
