+++ src/Pages/Settings.js
@@ -1,9 +1,21 @@
 import styled from "styled-components";
+import { useContext } from "react";
+import { AuthenticationContext } from "../Contexts/AuthenticationContenxt";
 
 function Settings() {
+  const { user, dispatch } = useContext(AuthenticationContext);
+
+  const onClickHandler = () => {
+    localStorage.removeItem("user");
+    dispatch({ type: "SIGNOUT" });
+  };
+
   return (
     <EditContainer>
-      <h1>Settings page</h1>
+      <h1>{user.email}</h1>
+      <StyledButton type="button" onClick={onClickHandler}>
+        Signout
+      </StyledButton>
     </EditContainer>
   );
 }
@@ -16,5 +28,27 @@ const EditContainer = styled.div`
   justify-content: center;
   align-items: center;
   overflow-y: scroll;
+
+  
+  h1 {
+    margin-bottom: 2rem;
+  }
 `;
+
+const StyledButton = styled.button`
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  background-color: #6ee7b7;
+  border-radius: 10px;
+  padding: 1rem;
+  border: unset;
+  font-weight: bolder;
+  font-size: 1rem;
+  cursor: pointer;
+  &:hover {
+    background-color: #34d399;
+  }
+`;
+
 export default Settings;
